feat(Threads): Hide Ads
