--- patches/src/main/kotlin/app/revanced/patches/instagram/ads/Fingerprints.kt
@@ -1,13 +1,0 @@
-package app.revanced.patches.instagram.ads
-
-import app.revanced.patcher.fingerprint
-import com.android.tools.smali.dexlib2.AccessFlags
-
-internal val adInjectorFingerprint = fingerprint {
-    accessFlags(AccessFlags.PRIVATE)
-    returns("Z")
-    parameters("L", "L")
-    strings(
-        "SponsoredContentController.insertItem",
-    )
-}
