+++ patches/src/main/kotlin/app/revanced/patches/meta/ads/Fingerprints.kt
@@ -1,0 +1,13 @@
+package app.revanced.patches.meta.ads
+
+import app.revanced.patcher.fingerprint
+import com.android.tools.smali.dexlib2.AccessFlags
+
+internal val adInjectorFingerprint = fingerprint {
+    accessFlags(AccessFlags.PRIVATE)
+    returns("Z")
+    parameters("L", "L")
+    strings(
+        "SponsoredContentController.insertItem",
+    )
+}
